Rename ceph profiles to cloudceph
We are about to deploy a new ceph cluster in WMF and we would like to
make optimal use of the existing puppet module for ceph to manage both
the new and existing clusters.

The existing clusters that we have are managed by the WMCS team and the
servers within these clusters use the hostname prefixes 'cloudcephmon'
and 'cloudcephosd'.

The puppet profiles are currently called 'profile::ceph::mon' and
'profile::ceph::osd'.

Whilst it would be ideal to be able to use these profiles on both new
and existing clusters, they are very much tailored to the cloudceph
clusters and significant work would be required to make them applicable
generically.

Therefore the purpose of this CR is to rename the existing ceph
profiles so that they match the hostnames of the servers and make it
easier to develop new profiles to deploy the new cluster.

This change should therefore be a no-op as far as any configuration is
concerned. However, there is a change required in the private repo in
order to support the rename.

Here is the corresponding change to the labs/private repo:
https://gerrit.wikimedia.org/r/c/labs/private/+/880949

Bug: T326945
Change-Id: I68db625db739b26cc27d0ad20aa957d73809cdf6