puppet_compiler: serve pson.gz as application/json
The compilations at https://puppet-compiler.wmflabs.org/ have the full
PSON catalogues compressed with gzip. The Nginx frontend sets the header
`Content-Encoding: gzip` and the browser thus happilly display the file
inline rather than prompting to download it.

PSON is pretty much JSON and on my browser it would be pretty formatted,
syntax highlighted and with folding which makes it way easier to
navigate but it is received with no content type and the feature does
not trigger.

Set the header `Content-Type: application/json` when emitting a
`.pson.gz` file.

We have used a similar hint on the CI Jenkins when emitting compressed
artifacts such as compressed svg files or large JSON performance
reports ( T249268#6959834 ).

Hosts: O:puppet_compiler::puppetdb
Change-Id: Ic6268ce90f2b646a3e830d0de328464852fe0d77