gitlab: exclude shell scripts and other backups from rsync jobs
This change excludes shell scripts and other backup types from backup
rsync jobs. The rsync jobs use --delete flag to make sure old backups
are removed on replica machines. This flag causes the backup/restore scripts
and config backup in the same folder to be deleted too. With the new exclude
parameter this files should remain in the folder on the replica.

Bug: T274463
Bug: T326315
Change-Id: I44ffee451b6b65fc73b99f30e230863ed7028527