role::kafka::jumbo::broker: update firewall rules for centrallog1001
For some reason we were not able to put the AAAA DNS resolution
for centrallog1001 in the Kafka Jumbo's firewall rules at the time.
After checking the note with Andrea we realized that the comment
is probably stale, since a AAAA record is present for
centrallog1001 nowadays. It seems less confusing and error prone
to avoid a raw IP, hence this change.

Change-Id: I72a5d5cd3a45ca9272773e049d018af89bca4ee6