gerrit: listen on all ports, DROP requests to host
c6ce0404c06 made Apache 2 to only listen on the Gerrit service IPs. The
intent is to prevent serving requests made to the hostname such as
https://gerrit1001.wikimedia.org/ which does not match Gerrit
configuration and does not have a valid TLS certificate.

When rebooting the server, Apache 2 is unable to start because the IPv6
address is not ready yet. I have tried to fix it by having the apache2
service to require `network-online.target` but that does not wait for
the IPv6 address to be fully configured. The target is reached while
IPv6 is still running the duplicate address detection (DAD) which takes
a few second to achieve. It is only after the detection has completed
that the IPv6 is fully ready and listenable on.

Since the intent was to not serve any request on the host IP:

- adjust the Gerrit HTTP ferm rule to only allow requests made to the
  service IPs (`$ipv4` and `$ipv6`)
- remove the custom Apache ports configuration and stop removing the
  default ports configuration which comes with the `httpd` class.

I am keeping the `wait_network_online` albeit it might not be necessary
anymore but it doesn't hurt to keep it around.

This should address Apache not starting up after the server is rebooted.

This reverts commit c6ce0404c0668abc6b8f80bf5df862fa61968a42

Bug: T326125
Hosts: O:gerrit
Change-Id: If56233338332542b751f263226c01adef63176c7