esitest: remove deprecated nbproc config option
On remaging cp2027 to bullseye, it was observed that the esitest service
fails to start because of a deprecated config in haproxy, nbproc.

    nbproc is not supported any more since HAProxy 2.5. Threads will
    automatically be used on multi-processor machines if available.

We are using two versions of haproxy in production: haproxy 2.4 and 2.6.
In 2.4, nbproc was already deprecated and assumed to be set to 1 so
let's remove it entirely.

Bug: T321309
Hosts: cp2027.codfw.wmnet, cp2028.codfw.wmnet, cp2033.codfw.wmnet
Change-Id: I78d5bf60a9edef5a78e7b94c2321e6cd3c4b15f2