phabricator: ensure phd uid/gid can not be changed
The `phd` uid is defined by `modules/admin/data/data.yaml` and there is
no need to be able to change it via a Puppet class parameter.

That was my original intent in the change to dedupe the phd user
creation (42e4358f6c) but it got amended and the uid/gid were made
parameters.

- remove uid/gid phabricator::phd::user class parameters
- assert the values are of type `Admin::UID::System::Global`

Bug: T326146
Hosts: O:phabricator
Change-Id: Id343c28c07840655723481e3e57e84422f8fc2e2