P:wmcs::services: adjust toolsdb pinning
Instead of pinning to a specific version (that might not be available
after an updated version has been made available), set the entire
package as priority 90. According to apt_preferences(5), packages with
priority between 0 and 100 will be installed but won't be upgraded,
which is exactly what we want here and what was the original intention
of the pin rule.

Not touching the equivalent 10.1 rule as that server is going to go
away soon.

Change-Id: I84a79a76813f266cb22420e55b58d70b316c9f93