Rabbitmq: use OpenStack bpo packages for rabbit
Previously we were using the distro-specific package in eqiad
and the bpo package in codfw1dev. This should get us to the bpo
package on both deployments.

My hope is that there are bugfixes in the bpo that will solve
some issues.

For the current version (zed) this will upgrade us from
3.8.9-3 (Bullseye) to 3.9.13-1~bpo11+1 (zed)

Bug: T328155
Change-Id: I0defdf41e1266382199dd1842c75175f70a047d5