Rotate aphlict logs either daily, or when they reach 1G
This rotates aphlict logs when they hit maxsize = 1G, or daily,
whichever comes first. It also keeps more of them, since we will be
rotating more frequently.

This requires running the logrotate script more frequently, so we add
a systemd timer to take care of this. We don't disable the existing
systemd timer (installed by the debian package) which runs logrotate
daily.

Bug: T325246
Change-Id: I8ee5e44b6c58eeb21d8dec808a8bc8607276c240