configmaster: Remove disc_desired_state.py
The list of services and their status is not kept up to date, and the
script is broken because of a non-working confctl command.
The related alarm has been in WARN for 2 months, causing me to think it
is not needed and contributes to alarm fatigue.

Bug: T327663
Hosts: alert1001.wikimedia.org, puppetmaster1001.eqiad.wmnet
Change-Id: I4dbde265be1c7c49dab832869f84d3589efbef54