Configure the new ceph servers with mon and mgr daemons
This is a role and a set of profiles to begin configuring the new Ceph
cluster. The intention is that these new profiles in the profile::ceph
namespace should ideally be applicable to any ceph cluster in future.

Consequently, their inspiration is the existing cloudceph profiles and
the intention is to minimise the necessary changes to the underlying
ceph module.

This commit adds the mon and mgr daemons to the new ceph servers and
configures them with a firwall permitting only the ceph servers
themselves to connect at the moment. This will be sufficient to
bootstrap the cluster, but we will need to add more rules for RBD
clients before it is usable.

They should be a noop operation for any of the servers in the WMCS ceph
clusters, which use the prefix of cloudceph in their hostnames.

Bug: T328123
Change-Id: I4458f9730d20bfd5036bd40e78809ae5c91055aa