phorge: git clone arcanist also from we.phorge.it, not Phacility
Turns out in one installation guide Arcanist, as a requirement
for Phabricator, is still cloned from Phacility's github account
but in another variant it's also cloned from we.phorge.it.

We prefer just relying on the single new entity of course since
the whole point of Phorge is that Phacility stopped being maintained.

Bug: T328595
Change-Id: I041319135d21673ade12de3fececd96d90890638