ores: add per-model metrics and fix label for response codes
ORES doesn't support Prometheus natively: it pushes metrics
via statsd (locally) and a prometheus-statsd-exporter is
in charge to collect them and expose their aggregation
in the Prometheus format. The ML team is in the process
of figuring out what traffic Lift Wing (the replacement of ORES)
will need to sustain, and we'd really need to have per-model
metrics rather than high level aggregation.

To make an example, here's the tcpdump from ores1001:
ores.ores1001.score_processed.wikidatawiki:64.778328|ms
ores.ores1001.score_processed.wikidatawiki.itemquality:64.778328|ms
ores.ores1001.score_processed.wikidatawiki.itemtopic:64.778328|ms

In the current case, we match only the first metric (the aggregation)
and we discard the other two, that would give us a lot of info
about what combinations of wiki/model receive the most traffic
in ORES (and hence what pods in Lift Wing will do as well).

Since the new metrics are per-wiki ones, and ORES supports
a lot of them, there will be an increase in metrics retrieved
and stored on the Prometheus masters, but the volume
shouldn't be massive.

Bug: T325763
Change-Id: I592b7ef6c95117bac3192824cb7ad118332d9d2c