Remove the GPU configuration from an-worker109[6-9]
These four hosts no longer have GPUs fitted, since all four cards
have been moved to dse-k8s-worker100[1-2].

This change will not reverse the configuration items made due to the
application of profile::python38 and the amd_rocm classes, so for
completeness we should reimage these nodes once the change is merged.

However, we are about to start upgrading the Hadoop nodes to bullseye
anyway, so perhaps we could roll this reimaging into that effort.

Hosts: an-worker1096.eqiad.wmnet,an-worker1095.eqiad.wmnet
Bug: T318696
Change-Id: I9c719d7d156740fd7bd53243972c532cecd433ad