Update the kubectl config files generated for the dse-k8s cluster
We no longer need to generate infrastructure use tokens and kubectl
files for spark-operator and spark-operator-deploy on the dse-k8s
cluster since we will be deployingthis service with the admin token.

However, we do want to generate kubectl configuration files for the
spark and spark-deploy infrastructure users. In time we will want these
tokens to be available on the stat hosts, but for now we will create
them with the others on the deployment hosts.

Also the cfssl-issuer tokens are not needed so are to be removed.

Bug: T322635
Bug: T318926
Change-Id: I244b276cb30da4d5612ec28b0c93f3c77b504489