prometheus::node_exporter: remove /var/lib/docker from ignored_mount_points
This change enables metrics for /var/lib/docker mount points in prometheus
node exporter. Metrics for that moint point are needed for dashboards and
alerts for gitlab-runners.
The node_exporter config was tested successfully in WMCS in T328972 without
seeing the errors mentioned in older tasks (like T211810 and T184469).

Bug: T328972
Bug: T329286
Change-Id: I06f578e23689c29be78eb888f1a8bbbf60b249f9