analytics::refinery::job::druid_load.pp: remove absented jobs
These jobs have been already migrated successfully to Airflow,
and are in production. We can remove them from here, since they
have been previously absented.

Actually, one of the removed jobs was absented for a long time,
but we can get rid of it too, since it's not used any more.

Bug: T328933
Change-Id: I6de7dece302d50192ca324ecd9dd6e030908dadb