dnsrecursor: update template to prepare for bullseye upgrade
In preparation for the upgrade of the hosts using the dnsrecursor module
where most hosts will be running buster but the test transition host
will be running bullseye (before all other hosts are upgraded to
bullseye), ensure that the pdns-recursor template remains compatible
with both buster and bullseye during this transition.

There is the added complexity of the pdns-recursor component as well but
since that will be always most the recent version of pdns-recursor (used
by Wikidough), we should be good on that.

PCC should be a NOOP on this for all C:dnsrecursor, except cloudservices
which is running pdns-recursor 4.6.2 running on bullseye.

Bug: T321309
Hosts: doh1001.wikimedia.org, dns1001.wikimedia.org, cloudservices1004.wikimedia.org
Change-Id: I1244450d13a6b6eadbe2ee288387d8bf83fde678