phabricator: stop/disable/mask phd based on phabricator_server setting
Instead of using the extra parameters to decide whether phd should
be running/enabled/unmasked or the opposite, base it (again) on the
central "phabricator_server" setting Hiera.

So only run 'phd' on the ONE active server and disable it on all
other servers.

This simplifies code and switching to a different server. Only one
place has to be edited and not multiple.

Previously we were hoping phd could run on a read-only DB and wanted
to be able to control these things independently, also for migrating
to new hardware. We most likely won't ever do that again though.

There is more, like whether dumps are enabled or not that could also
be changed to use this single setting.

Bug: T329285
Change-Id: I2fbb2f8405a1c30035c20f12e62f6bc2445a19d5