P:wmcs::services::toolsdb_replica_cnf: don't manage the directories
Do not change the directory permissions to www-data:www-data. This broke
PAWS, and is generally not wanted here.

If the purpose was to let the API edit the files, that would not have
worked anyways as for Toolforge they will be owned by the tool/user
account anyways, and for PAWS they will be owned by the tools.paws user.

You might want to consider using the CAP_DAC_OVERRIDE capability instead
of running the daemon as root.

Bug: T329377
Change-Id: Ib06e256569482036170b911aa5780c70cb5c2a5c