postgresql::user: Check some password is set for a user
With the current puppet policy we first create the user and then set the
password as long as the password is not set.  on bookworm Currently we
just check that a user exists to determine if the password is correctly
set.  This is in valid as the user will exist but with a blank password.
As such update the check to ensure we check that some password with
SCRAM-SHA-256 has been set

Hosts: auto
Change-Id: Idf380de20a5d40ae8c03dfd10a348f1bce9351ee