Add binder to the kernel module block list
Binder is an IPC driver used in Android, but also enabled in the Debian
kernel builds (e.g. to support the use of the Android debug bridge).

We don't need/use Binder and it was affected by a recent Linux
privilege escalation (CVE-2023-20928):

https://bugs.chromium.org/p/project-zero/issues/detail?id=2374

And also other vulnerabilities in the past (CVE-2019-2215 CVE-2019-2214).

Add it to the modprobe block list as defense in depth to prevent it
getting loaded indirectly (via package postinst or similar).

Change-Id: Icb307b2ad744ccfba265e52f34ed18af2d5ac31a