Try libmariadb-java with sqoop on bullseye
For the buster upgrade we encountered an issue with using
libmariadb-java with sqoop. The result of that was that we
forward-ported libmysql-java to buster and used that package.

Rather than forward-port that old package again, we would like to try
once again to use the libmariadb-java version that is supplied, but this
time the bullseye version.

This change will require a dependent change to the analytics-refinery
sqoop-mediawiki-tables job to select the mariadb JDBC driver.

Bug: T329363
Bug: T278424
Change-Id: Ib0b96f0b26ecf5f576405a5a2ecf60d891eb42a4