simplelamp2: change default mariadb datadir to /var/lib/mysql/
If used on a cloud VPS, do not use /srv/ as the default datadir.

But do not change the default in the profile because that is also
used by some things in production, though only a few of them, and
not the main prod DBs, which use other packages.

To be able to do a Hiera lookup to override the new default
for a few exising projects, refactor to turn the role into a role
and a profile.

Bug: T329571
Change-Id: I8b1e29dca6f2adc9bc02ad4bce343d7c8e73b79e