serviceops-collab: switch alert severity to 'task' globally
Use the new "severity"-level added in I11123889de00a104 that
only creates tickets for all existing alerts mapped to serviceops-collab.

Most were "warning" and for gitlab we give up on making a difference
between primary and other servers.

Though we keep the option to change that again some day.

Bug: T329587
Change-Id: I32f35d5a9dfdd4bc2cd033ebff1166db8fdb0d8f