gitlab_runner: Set pull_policy = ["always", "if-not-present"] on WMCS runners
This will try to "always" pull the image, but if it fails (e.g.
Docker Hub rate limits), it will use an already-pulled image
if it exists for WMCS runners.

The default is already "always", so this is just adding
"if-not-present" as a fallback for WMCS runners. Trusted Runners
will stay on "always".

See <https://docs.gitlab.com/runner/executors/docker.html#set-multiple-pull-policies>.

Bug: T329216
Co-Authored-by: Jelto <jwodstrcil@wikimedia.org>
Change-Id: Iac375d01ffd1f7c0e758d0b1040802b0fbb05b54