Adjust interface names for cloudcephosd1001 and cloudcephosd1002
Cloudcephosd nodes make adjustments to /etc/network/interfaces on
top of what the default NIC configuration adds at install time.

The reason for this is to enable a second port on the box.  This
config needs to know the linux device name to configure.

Unfortunately the upgrade from buster to bullseye (and firmware
upgrade of NIC to allow this possibly) means that the NIC names
have changed as part of the upgrade.

This patch should ensure the hosts are configured with the correct
interface names.

Bug: T329498
Change-Id: I633ee2c75a108b0ce17c6891c6e240828c739569