webperf: Remove broken HeaderName/ReadmeName for arclamp file listing
These haven't worked for several years because these files exist on
the webperf server that serves performance.wikimedia.org, and not
on the backend arclamp server that serves the /arclamp directtory.

The other references here (to CSS and PNG files) also don't exist
on the arclamp server, but that's fine because those are interpreted
as URLs and only need to work in the final product as served from
performance.wikimedia.org, whereas the other ones are interpreted
as server-side includes.

The unused files are removed in performance/docroot.git by
change I00820f66385d69d77b1c09a98d2f1a35f4ee2f05.

Change-Id: I8d392a6b74c2365ec0785529b2fe60e5c2eb1e37