Remove the ores::base class from the analytics cluster
We have determined that noone is using this functionality on the hadoop
cluster and the stats servers, we can remove it and the git::lfs class
that it required.

Note that there is still some incompatibility between base::ores and
bullseye, so where this class is deployed to other servers will still
cause issues upgrading. One example is the replacement of the enchant
package with enchant-2 which may cause issues for tools dependent on
that wrapper.

Bug: T329363
Change-Id: Id387c41440d894f9827106fe3e6b035d1c75a75e