dnsrecursor: enable webserver for bullseye installation of pdns-rec
The dnsrecursor module by default disables the pdns-rec webserver that
also exposes the /metrics endpoint. This feature is instead provided by
profile::prometheus::pdns_rec_exporter for the version in buster where
the version of pdns-rec does not export Prometheus metrics.

In preparation for the upgrade to bullseye, switch to the metrics
endpoint provided by pdns-rec itself instead of using
prometheus-pdns-rec-exporter.

For a related commit that sets the installation of
prometheus-pdns-rec-exporter to only buster, see I9db5369.

Additionally, make the webserver port configurable to maintain
compatibility with prometheus-pdns-rec-exporter (port 9199) and
cloudservices (default port 8082).

Bug: T321309
Hosts: auto
Change-Id: I4e8f9eb110544748f3f06901b26a4f3d81e1dee1