gitlab: remove dedicated restore logfile and log to syslog only
This change removes the dedicated restore log at /var/log/gitlab-restore-backup.log.
The log file is not rotated or collected by rsyslog and around 0.5GB in size. Also
some logs (stdout) are in the log file and some logs (stderr) in syslog/journal.

This change aligns the logging so that all logs end up in syslog/journal and should
be found using journalctl -u backup-restore.service.

Bug: T326315
Change-Id: I14899c8146d0f5554b539e6188e433bc504ca0a9