P:dns::auth::update: add safe.directory for gdnsd bootstrapping
During the reimaging of dns4004, authdns-local-update failed to run and
because of that gdnsd was failing to start with the right configuration.

When executing:

    utils/deploy-check.py -g /srv/git/netbox_dns_snippets --deploy --no-gdnsd

It was returning the error:

    raise ValueError("SHA could not be resolved, git returned: %r" %
    (header_line.strip()))

This results from the update to Git, whereby the root user cannot read
the git directories owned by the authdns and netboxdns user. To fix
this, we use the safe directory directive, as outlined in CR
888053 to allow authdns-local-update to proceed.

Bug: T321309
Hosts: dns1001.wikimedia.org, dns4004.wikimedia.org
Change-Id: I5934d18c566984071ac4a5d894644e946308b0f8