Add a postgresql database and user for airflow_search_platform
The Search Platform team is deploying a new Airflow instance which will
use version 2.5. As such it will not be compatible with MariaDB and will
require a PostgreSQL database. The password has already been added to
the private repository, so this change causes the database and its
grants to be created on the PostegreSQL server.

Bug: T326193
Bug: T327970
Change-Id: I4ba7571698a742ca5d7a9aeaa2297b7ea1ac6857