P:toolforge: k8s: haproxy: increase max connections
Increase the amount of max connections per ingress server from 250 to
1000, as we've been hitting the limit every couple of hours for a while.
(To check this yourself, query tools-prometheus for
"haproxy_server_current_sessions{backend="k8s-ingress"}".)

As with the old limit, the new limit is picked fairly arbitrarily. Based
on the resource usage by the ingress pods we'll be fine with the new
limits, and we can always look into tuning the ingress pods if we start
seeing issues there.

Change-Id: Ic3ce27e8b41ae974dbaa9dc55f80353bd928d1e0