kubernetes: Continue to use the cergen cert for service-account signing
kube-apiserver and controller-manager rely on the service-account
certificate to be shared between all control-plane nodes. As there is
currently no easy and reliably way to to that with PKI, fall back to
using cergen cert (profile::kubernetes::master::service_cert) for that
until we have a proper solution in place.

Hosts: auto
Bug: T329826
Change-Id: Ic635ec85d48abcb0cb06384833f1155071b68981