fix(presto): create pkcs12 server file with intermediate certificate
Newly created server.p12 file is missing the intermediate certificate
on an-test-presto1001.eqiad.wmnet. Thereis the same pattern on an-test-coord1001

On prod cluster the intermediate certificate is well in the file

From profile::pki::get_cert the file containing the intermediate certificate
is return in the chained key of the path hash.

Bug: T329361
Change-Id: I3babab1a3b13e8c6ecb78f4db72b00dd27752f77