replica_cnf_api_test: check if user with id USER_ID exists
check whether user with id USER_ID exists and if not
use USER_ID for regex testing for file permissions and user.
If the user exists use the target username for regex
check or test will fail.

This is neccessary because we are testing replica_cnf on
two different vms (stretch and bullseye) and the test
user doesn't exist it labstore1004 but does in dbusers-nfs1.

Bug: T303663
Change-Id: I1be87cfba01d0b2449ea3d5ec9d050180cf71501