mediawiki-cache-warmup: Rewrite in Python
This replaces the Node script with a Python equivalent. The
functionality is mostly the same, and they read the same URL files.
Some minor differences:

- https:// URLs are no longer converted to http:// plus an
  X-Forwarded-Proto header. Almost all traffic to the app servers is now
  encrypted, and load balancing (for "spread" mode) is no longer
  available on port 80, anyway.

- The confctl command to determine target hosts in "clone" mode now
  uses service=nginx, since service=apache2 was removed in
  Ic908932f3c39e75e6e94931d8c2032427f5c6465.

- The "clone-debug" mode (sending requests out to the edge, with
  X-Wikimedia-Debug set) isn't ported. Since all traffic is HTTPS,
  `spread mwdebugNNNN.eqiad.wmnet` is functionally equivalent now, and
  we can add a more natural synonym if we end up using it, but I suspect
  it's not needed.

A followup change will update the warmup cookbook to run this script.

We may eventually move this into spicerack (in which case we'd also use
spicerack.confctl, etc.) but we'll consider that change separately.

Bug: T288867
Change-Id: I72b198e5de4ed9781bedd5718b312331a7c26b4a