Extend profile::nginx with support for new Nginx packaging layout
Starting with the version of Nginx in Bookworm, nginx has switched
away from the previous (and somewhat random) selection of modules
built into nginx-light, nginx-full and nginx-extras.

Instead there's now a single nginx binary package and the specific
modules needed can be installed as libnginx-mod-foo as plugins.
(In Bookworm -light, -extras and -full still exist as transition
packages).

Extend the variant flavour and add a new Hiera argument to pass
a specific (if needed) list of modules to install.

Also move to ensure_packages while we're at it.

Hosts: C:nginx
Bug: T329529
Change-Id: I88690f57e61855e0406c0621caea058539e483f5