doc: fix hostname used in http::blackbox monitoring
We are seeing in logstash:

x509: certificate is valid for doc.discovery.wmnet...
..not doc.wikimedia.org-doc2001.codfw.wmnet

The resource name must be based on the host name to avoid
puppet duplicate declaration error.

We are going to check on both backends, doc1001 and doc2001.

Set server_name explicitly, don't let it default to $title.

Set instance_name to $other_host to be used as label.

We really want to check the same virtual host on both
instances.

Bug: T327973
Bug: T330233
Change-Id: I162c7758abe4433efba3d750e2092da9cfe1dc74