Deploy analytics-refinery to search airflow instance
The search airflow instance invokes the refinery python module,
specifically the refinery-drop-older-than.py script. As this
python module isn't packaged in a way that can be installed into
a deployed artifact and used in the typical manner via airflow,
ship the git repo to the instance so it can invoke the script
directly.

Bug: T329870
Change-Id: I99bb40e9b5ac175538b68d3bc3e66c66ab930f96