ci: set port 1443 for https monitoring
On contint hosts, there is a webserver on port 80 and an
envoyproxy on port 1443 that terminates TLS.

None of them are listening on default port 443.

In logstash, even after opening firewall holes for prometheus
hosts, we still see ":443: connect: connection refused".

[contint1002:~] $ telnet localhost 443
Trying ::1...
Trying 127.0.0.1...
telnet: Unable to connect to remote host: Connection refused

[contint1002:~] $ telnet localhost 1443
Trying ::1...
Trying 127.0.0.1...
Connected to localhost.
Escape character is '^]'.

Bug: T327972
Change-Id: Ibb619904ee64a643a4f6b84c8ac92a8489c2caa4