peopleweb: switch rsync source and dest between eqiad and codfw
This changes which host is the source and which host is the destination
when it comes to (manual) rsync of data between people* hosts.

It changes firewall rules and rsyncd config but the actual sync commnd
is still run manually.

It also smartly changes the MOTD, like we do on deployment servers, so users
who ssh to one of them will be told either that is the active server
or the wrong server to upload files to.

I can also do this anytime by myself or coordinated in some way
with the rest of the DC switch-over.

After this is done the DNS change to actually switch which
backend is used by ATS is separate at I073e70e18ad727.

Bug: T330091
Change-Id: I8d4ec62d508eaaf4827703e2446a446214c5bf54