role::etcd::v3::kubernetes::staging: move certs to PKI
The absence of a TLS SAN causes some errors when the etcd daemons
in the ensemble try to contact each other, causing a ton of spam
in syslog and related logs. The move to PKI fixes this since
the SAN is added automatically for bullseye nodes.

The wikikube codfw cluster is already using PKI, as well as aux
and dse.

Bug: T329717
Change-Id: I624c9e5bf6ca3aca91486847061b6d168125e017