Ensure that the airflow database names match existing conventions
The initial databases and users were created using WMF team names and as
such the analytics component was renamed to data_engineering. However,
the existing convention was to use the name of the airflow instances,
which had already been set to analytics and analytics_test.

This commit corrects that error and creates new databases and users to
ease the migration from Airflow 2.1 and MariaDB to Airflow 2.5 and
PostgreSQL.

Bug: T319440
Bug: T326193
Change-Id: I026b4b21a0e57a3493da456b8437458c5e74b389