Blacklist f2fs
This is a filesystem we don't use on our servers (and won't ever need)
and which sees little upstream attention, there are two security issues
dating back to 2019, which are still unfixed to this date (CVE-2019-19449
and CVE-2019-19814).

Blacklist as defense in depth to reduce attack surface (e.g against container
images potentially using it). I'll be adding more in the future, but
let's already add f2fs due to the unfixed issue listed above.

Change-Id: I92f499302131194eb96d3a41fb2ef66695e6c32a