Split out a new class for phabricator::config
While trying to do a scap deploy of phabricator on aphlict hosts, we
ran into a problem of the phabricator configs being required, but not
present.

The solution is to load the phabricator configs onto the aphlict hosts.
The simpler option is to split this out into a separate class that can
be added to the aphlict class as needed

Change-Id: I550a7bad85848c4455d6b5fa8a3ed84560ece0a4