switch os-reports.wikimedia.org from eqiad to codfw
The files served by https://os-reports.wikimedia.org
are rsynced over from a cumin host where they are generated.

A timer/service to sync them exists and works on both backends
and both have today's report as the latest one.

So this can also just be switched over to codfw.

Bug: T330090
Change-Id: I86c1407e05d44781cf685e803a79a6588589599b