gitlab: enable restore for replicas, disable on active_host
This change enables and disables the restore depending on the active/
replica state of the instance. For the currently active host, restore
is disabled and for replicas restore is enabled.
In the past configuration of the hiera value was done manually and forgotten,
which lead to data loss and restore on the active host.

Bug: T329931
Bug: T330717
Change-Id: Icdba94fb61bef1bff8ecde7d483faab4c0723829