systemd::timer::job: fix email body indentation
Upon receiving the notification email, the body starts with an empty
newline and the message is indented after the Python source code
indentation.

The reason is `textwrap.dedent` removes indentation based on the first
whitespace it encounters (space, tabulation or **newline**).  The triple
quoted string used to forge the body starts with a newline:

    """\n
    Systemd timer ran the following command:\n

The newline is picked up by dedent which then does not remove leading
spaces. From the Python example:

    def test():
        # end first line with \ to avoid the empty line!
        s = '''\
        hello
          world
        '''

Add an antislash to prevent the newline character and allow dedentation
of ther rest of the text.

Since the command output might use indentation for display purpose, move
`textwrap.dedent` call to apply to the formatting template rather its
expansion which will preserve the indentation from the command output.

Follow up 5ad5de58c.

Hosts: basic
Bug: T330120
Change-Id: Ia1083d301df754b72f04036d9c86c5e5b0fc4189