elastic relforge: rm rules for Jenkins
We had the idea of shipping Jenkins builds output to an ElasticSearch
cluster. It was proposed to use the relforge cluster and we thus added
a firewall rule to permit Jenkins host to write to it.

Since the project (T78705) never realized, remove the unused rules.

This reverts commit e1f31a4179128b8f2f4b1c8cd76d294ebc03aced.

Hosts: relforge1003.eqiad.wmnet, contint1002.wikimedia.org
Bug: T78705
Change-Id: I92d4c7bcfecf8a00342d3d76d07f401973619ffb