contint: manage dsh target from Puppet DB
This will ease future maintenance as we add/remove hosts and ensure the
dsh groups are matching the actual deployments.

scap requires files holding the list of target hosts which we are
manually managing as a Hiera list. Since targets have `scap::target`
applied and it is collected in the PuppetDB, we can have Hiera to
retrieve the list of targets for us.

The pattern has been introduced by Giuseppe for the k8s-workers in
change Ic53709f2fb55f532d9162541a274d545cb3c29a9.

The `jenkins-ci` and `jenkins-rel` dsh groups will be addressed in
follow up changes, they share the same repository but have different
roles.

Change-Id: Ib182162497a17f329e4948a15647d49b355a4e7d