OpenStack: collapse 'user' OpenStack role into 'reader' role
Other than ssh + bastion access these roles have had the same
meaning for quite a while.

This patch adds ssh access for anyone with the 'reader' role on a
project, and eliminates the 'user' role.

Note that service users e.g. 'novaadmin' might have the 'reader'
role on a domain rather than on a project; this provides read access
but not ssh access to subprojects.

Bug: T330759
Change-Id: Id8ba428992619674324b6ea50ef8958c7c18899b