profile::logstash: improve the ORES filter
The ORES filter is currently not catching some traffic due to
the IP field, that in some cases could be a list of IPs separated
by a comma (I've always seen max 2 of them in the ORES access
logs though). Since the first one is the most important,
it is better to discard the rest of the data.

Add also a specific test case for the new version of the filter.

Bug: T325763
Change-Id: I56ec3d3a464e6d94409a6332c1a8fc9cf4dc200d