kerberos: update the motd so its not to big
he motd messages are starting to get so big that they scroll of the
screen and can be easily missed.  as example the apt1001 the motd on
apt1001 is not 46 lines long and the important information that it is not
the active server is completely missed on my standard laptop terminal
window size[1]

This Cr updates the MOTD so its smaller but uses color to try and
highlight the important information.

[1]https://pasteboard.co/Uih3UsTEFuae.png

Hosts: O:apt_repo
Change-Id: I02423fc3da43dfcd2e06d17e12025d0f20a76293