apt: update the motd so its not to big
The motd messages are starting to get so big that they scroll of the
screen and can be easily missed.  as example the motd on apt1001 is
now 46 lines long and the important information that it is not
the active server is completely missed on my laptop standard terminal
window size[1]

This Cr updates the MOTD so its smaller but uses color to try and
highlight the important information.

[1]https://pasteboard.co/Uih3UsTEFuae.png

Hosts: O:apt_repo
Change-Id: Ib6aa04f420d6954a2d574afb8201097e4c7e534b