ceph: Add puppet management of OSDs on new ceph cluster
This commit adds a defined type of ceph::osd along with a ceph::osds
class and a corresponding profile::ceph::osds manifest.

This allows us to use puppet to manage the configuration of the OSD
daemons, which has so far only been carried out with one or more
cookbooks.

The ceph::osds class makes use of the new ceph_disks fact that has been
deployed to our new cluster. The reason for doing this is that we wish
to move away from using standard device names (/dev/sd*) since these are
volatile and may change between reboots.

The new fact allows us to associate specific storage slots with a
drive's wwn and serial number. We use the wwn value to obtain its device
name at runtime. SAS and SATA drives behave differently in regard to
their wwn so this module tries to accommodate both interface types.

Certain slots can be permanently excluded from the ceph storage topology
by providing their slot identity via hiera. This is intended to make it
simple to identify the drives used by the O/S or similar.

For OSDs backed by spinning disks, optional support for using an NVMe
device to host the Bluestore DB is present. The ceph::osds class
partitions the device according to the number of HDDs in use.

OSDs may be removed from the cluster in order to support replacement.

Bug: T330151
Change-Id: I0957a23d492ec182f0cc46fa3dfee2a44288eee5