Add an entry in the service catalog for the aqs service running in codfw
This will cause the ip address of 10.2.1.12 to be added as a
realserver_ip to the aqs20[01-12] hosts. THese are currently failing to
run puppet because they are expecting this entry in the service catalog
to be present.

There are still some outstanding questions about whether to serve AQS
traffic from codfw, since one of its back-end data stores in in eqiad
only. However, adding this realserver_ip will not cause in itself pool
the servers and route traffic. The servers themselves are still marked
as inactive in confctl.

I will prepare a CR with a corresponding DNS record for
aqs.svc.codfw.wmnet

Bug: T331115
Change-Id: I947df1298ade13d36ce6a07745b152be502ff67c