prometheus: ensure we update the mtime if no update required
Currently when:
  * the assembled file matches the content of all fragments and
  * one (or more ) of the fragments mtime is greater then the assembled file

Puppet gets confused as it first runs:
  sudo prometheus-assemble-config --onlyif blackbox
to check if a change is needed, which returns true as the mtime of a
fragment is greater then the assmbled file.  So puppet calles
  sudo prometheus-assemble-config blackbox
which dose nothing as the content of the assembled file matches the
content of the fragments.

As such this Cr updates the code to touch the assembled file in this
situation so that it will not trigger on the next run of
  sudo prometheus-assemble-config --onlyif blackbox

I have also added a little bit of logging

Change-Id: I55feed0abe99e22c4798a287f648ab5006740dbe