ntp: remove peer xleave option
xleave is intended to increase the accuracy of timing measurements
between peers, as documented in:

 https://www.eecis.udel.edu/~mills/ntp/html/xleave.html

We've been using this option on our "peer" lines for our NTP
clustering for years, but at least since upgrading to bullseye's
version of ntpd, this seems to have broken our peering
associations, claiming "bad dispersion", and removing the option
seems to fix it.

Since having peering at all is better than having more-accurate
peering which fails to work, we'll go with this as the fix.

Change-Id: I88f31b02c3d955a5812bbf68df6dcb7a8c9199a1